hair_salon_bot/requirements.txt
python-telegram-bot==13.15
scikit-learn==1.2.2
joblib==1.2.0
